@@ -12,42 +12,53 @@
 // limitations under the License.
 // ----------------------------------------------------------------------------------
 
+using Microsoft.Azure.Commands.Common.Authentication;
+using Microsoft.Azure.Commands.Common.Authentication.Abstractions;
+using Microsoft.Azure.Commands.Common.Authentication.Models;
+using Microsoft.Azure.Commands.ResourceManager.Common;
+using Microsoft.Azure.Commands.TestFx.Mocks;
+using Microsoft.Azure.Commands.TestFx.Recorder;
+using Microsoft.Azure.ServiceManagement.Common.Models;
+using Microsoft.Azure.Test.HttpRecorder;
+using Microsoft.Rest.ClientRuntime.Azure.TestFramework;
+using Microsoft.VisualStudio.TestPlatform.ObjectModel;
+using Microsoft.WindowsAzure.Commands.ScenarioTest;
 using System;
 using System.Collections.Generic;
 using System.Diagnostics;
 using System.IO;
 using System.Linq;
+using System.Reflection;
 using System.Runtime.CompilerServices;
-using Microsoft.Azure.Commands.Common.Authentication;
-using Microsoft.Azure.Commands.Common.Authentication.Abstractions;
-using Microsoft.Azure.Commands.Common.Authentication.Models;
-using Microsoft.Azure.Commands.ResourceManager.Common;
-using Microsoft.Azure.ServiceManagement.Common.Models;
-using Microsoft.Azure.Test.HttpRecorder;
-using Microsoft.Rest.ClientRuntime.Azure.TestFramework;
-using Microsoft.WindowsAzure.Commands.ScenarioTest;
+using Xunit;
 using Xunit.Abstractions;
 
 namespace Microsoft.Azure.Commands.TestFx
 {
     public delegate IRecordMatcher RecordMatcherDelegate (bool ignoreResourcesClient, Dictionary<string, string> resourceProviders, Dictionary<string, string> userAgentsToIgnore);
-
+    
     public class TestManager : ITestRunnerFactory, ITestRunner
     {
-        private readonly string _callingClassName;
-        private string _projectSubfolderForTestsName = null;
-        private string _newPsScriptFilename = null;
-        private Dictionary<string, string> _matcherExtraUserAgentsToIgnore;
-        private Dictionary<string, string> _matcherNewUserAgentsToIgnore;
-        private Dictionary<string, string> _matcherResourceProviders;
-        private Action _mockContextAction;
         protected EnvironmentSetupHelper Helper;
-        protected readonly List<string> RmModules;
-        protected readonly List<string> CommonPsScripts = new List<string>();
-
-        protected RecordMatcherDelegate RecordMatcher { get; set; }
-
         protected XunitTracingInterceptor Logger { get; set; }
+
+        private readonly string _testProjectPath;
+        private readonly string _callingClassName;
+        private string _testBaseFolderName = String.Empty;
+        private string _psTestScript;
+        private readonly List<string> _psAzModules;
+        private readonly List<string> _psCommonScripts = new List<string>();
+
+        private RecordMatcherDelegate _recordMatcher;
+        private bool _matcherIgnoreResourceClient = true;
+        private Dictionary<string, string> _matcherUserAgentsToIgnore;
+        private Dictionary<string, string> _matcherResourceProvidersToIgnore;
+        
+        private Action _initAction;
+        private Action<MockContext> _mockContextAction;
+        private Action _cleanupAction;
+
+        private Func<MockContext, object>[] _initializedServiceClients;
 
         /// <summary>
         /// Factory method
@@ -60,29 +71,26 @@ namespace Microsoft.Azure.Commands.TestFx
             var callingClassName = string.IsNullOrEmpty(callerFilePath)
                 ? null
                 : Path.GetFileNameWithoutExtension(callerFilePath);
-            return new TestManager(callingClassName).WithTestOutputHelper(output);
+
+            return new TestManager(callingClassName, Assembly.GetCallingAssembly()).WithTestOutputHelper(output);
         }
 
         /// <summary>
         /// ctor
         /// </summary>
         /// <param name="callingClassName"></param>
-        protected TestManager(string callingClassName)
+        protected TestManager(string callingClassName, Assembly callingAssembly)
         {
             Helper = new EnvironmentSetupHelper();
             _callingClassName = callingClassName;
+            _testProjectPath = GetTestProjectPath(callingAssembly);
 
-            RmModules = new List<string>
+            _psAzModules = new List<string>
             {
                 Helper.RMProfileModule,
                 Helper.RMResourceModule,
             };
-
-            RecordMatcher = (ignoreResourcesClient, resourceProviders, userAgentsToIgnore) =>
-                new PermissiveRecordMatcherWithApiExclusion(ignoreResourcesClient, resourceProviders, userAgentsToIgnore);
         }
-
-        #region Builder impl
 
         /// <summary>
         /// Sets a name of the subfolder where a test project keeps tests
@@ -91,18 +99,7 @@ namespace Microsoft.Azure.Commands.TestFx
         /// <returns>self</returns>
         public ITestRunnerFactory WithProjectSubfolderForTests(string folderName)
         {
-            _projectSubfolderForTestsName = folderName ?? "ScenarioTests";
-            return this;
-        }
-
-        /// <summary>
-        /// Add helper scripts
-        /// </summary>
-        /// <param name="psScriptList"></param>
-        /// <returns>self</returns>
-        public ITestRunnerFactory WithCommonPsScripts(string[] psScriptList)
-        {
-            CommonPsScripts.AddRange(psScriptList);
+            _testBaseFolderName = folderName ?? "ScenarioTests";
             return this;
         }
 
@@ -113,7 +110,31 @@ namespace Microsoft.Azure.Commands.TestFx
         /// <returns>self</returns>
         public ITestRunnerFactory WithNewPsScriptFilename(string psScriptName)
         {
-            _newPsScriptFilename = psScriptName;
+            _psTestScript = psScriptName;
+            return this;
+        }
+
+        /// <summary>
+        /// Add helper scripts
+        /// </summary>
+        /// <param name="psScriptList"></param>
+        /// <returns>self</returns>
+        public ITestRunnerFactory WithCommonPsScripts(string[] psScriptList)
+        {
+            _psCommonScripts.AddRange(psScriptList);
+            return this;
+        }
+
+        /// <summary>
+        /// Clears default RM modules list and sets a brand new
+        /// </summary>
+        /// <param name="buildModuleList"></param>
+        /// <returns></returns>
+        public ITestRunnerFactory WithNewRmModules(Func<EnvironmentSetupHelper, string[]> buildModuleList)
+        {
+            _psAzModules.Clear();
+            var moduleList = buildModuleList(Helper);
+            _psAzModules.AddRange(moduleList);
             return this;
         }
 
@@ -126,20 +147,7 @@ namespace Microsoft.Azure.Commands.TestFx
         public ITestRunnerFactory WithExtraRmModules(Func<EnvironmentSetupHelper, string[]> buildModuleList)
         {
             var moduleList = buildModuleList(Helper);
-            RmModules.AddRange(moduleList);
-            return this;
-        }
-
-        /// <summary>
-        /// Clears default RM modules list and sets a brand new
-        /// </summary>
-        /// <param name="buildModuleList"></param>
-        /// <returns></returns>
-        public ITestRunnerFactory WithNewRmModules(Func<EnvironmentSetupHelper, string[]> buildModuleList)
-        {
-            RmModules.Clear();
-            var moduleList = buildModuleList(Helper);
-            RmModules.AddRange(moduleList);
+            _psAzModules.AddRange(moduleList);
             return this;
         }
 
@@ -149,16 +157,29 @@ namespace Microsoft.Azure.Commands.TestFx
         /// <param name="userAgentsToIgnore">Dictionary [userAgent:apiVersion] to ignore</param>
         /// <param name="resourceProviders">Dictionary [resouceProvider:apiVersion] to match</param>
         /// <returns></returns>
-        public ITestRunnerFactory WithNewRecordMatcherArguments(Dictionary<string, string> userAgentsToIgnore, Dictionary<string, string> resourceProviders)
+        public ITestRunnerFactory WithNewRecordMatcherArguments(Dictionary<string, string> userAgentsToIgnore, Dictionary<string, string> resourceProviders, bool ignoreResourceClient = true)
         {
-            _matcherNewUserAgentsToIgnore = userAgentsToIgnore;
-            _matcherResourceProviders = resourceProviders;
+            _matcherIgnoreResourceClient = ignoreResourceClient;
+            _matcherUserAgentsToIgnore = userAgentsToIgnore;
+            _matcherResourceProvidersToIgnore = resourceProviders;
             return this;
         }
 
-        public ITestRunnerFactory WithMockContextAction(Action mockContextAction)
+        public ITestRunnerFactory WithInitAction(Action initAction)
+        {
+            _initAction = initAction;
+            return this;
+        }
+
+        public ITestRunnerFactory WithMockContextAction(Action<MockContext> mockContextAction)
         {
             _mockContextAction = mockContextAction;
+            return this;
+        }
+
+        public ITestRunnerFactory WithCleanupAction(Action cleanupAction)
+        {
+            _cleanupAction = cleanupAction;
             return this;
         }
 
@@ -169,21 +190,7 @@ namespace Microsoft.Azure.Commands.TestFx
         /// <returns>self</returns>
         public ITestRunnerFactory WithRecordMatcher(RecordMatcherDelegate recordMatcher)
         {
-            RecordMatcher = recordMatcher;
-            return this;
-        }
-
-        /// <summary>
-        ///
-        /// </summary>
-        /// <param name="userAgentsToIgnore">
-        /// Dictionary to store pairs: {user agent name, version-api to ignore}.
-        /// Initial pair is {"Microsoft.Azure.Management.Resources.ResourceManagementClient", "2016-02-01"}
-        /// </param>
-        /// <returns>self</returns>
-        public ITestRunnerFactory WithExtraUserAgentsToIgnore(Dictionary<string, string> userAgentsToIgnore)
-        {
-            _matcherExtraUserAgentsToIgnore = userAgentsToIgnore;
+            _recordMatcher = recordMatcher;
             return this;
         }
 
@@ -195,56 +202,90 @@ namespace Microsoft.Azure.Commands.TestFx
             return this;
         }
 
+        public ITestRunnerFactory WithManagementClients(params Func<MockContext, object>[] initializedManagementClients)
+        {
+            _initializedServiceClients = initializedManagementClients;
+            return this;
+        }
+
         public ITestRunner Build()
         {
             SetupSessionAndProfile();
-            SetupMockServerMatcher();
-            HttpMockServer.RecordsDirectory = Path.Combine(AppDomain.CurrentDomain.BaseDirectory, "SessionRecords");
+            SetupMockServer();
             Helper.SetupModules(AzureModule.AzureResourceManager, BuildModulesList());
             return this;
         }
 
         public void RunTestScript(params string[] scripts)
         {
-            var sf = new StackTrace().GetFrame(1);
-            var className = sf.GetMethod().ReflectedType?.ToString();
-            var methodName = sf.GetMethod().Name;
+            RunTestScriptCore(null, null, null, scripts);
+        }
 
-            using (var mockContext = MockContext.Start(className, methodName))
-            {
-                _mockContextAction?.Invoke();
-                AzureSession.Instance.ClientFactory = new TestClientFactory(mockContext);
-                Helper.SetupEnvironment(AzureModule.AzureResourceManager);
-                SetupAzureContext();
-                Helper.RunPowerShellTest(scripts);
-            }
+        public void RunTestScript(Action<MockContext> contextAction, params string[] scripts)
+        {
+            RunTestScriptCore(null, contextAction, null, scripts);
         }
 
         public void RunTestScript(Action setUp, Action tearDown, params string[] scripts)
         {
-            setUp?.Invoke();
-            RunTestScript(scripts);
-            tearDown?.Invoke();
+            RunTestScriptCore(setUp, null, tearDown, scripts);
         }
 
-        #endregion
+        public void RunTestScript(Action setUp, Action<MockContext> contextAction, Action tearDown, params string[] scripts)
+        {
+            RunTestScriptCore(setUp, contextAction, tearDown, scripts);
+        }
 
-        #region Helpers
+        private void RunTestScriptCore(Action setUp, Action<MockContext> contextAction, Action tearDown, params string[] scripts)
+        {
+            var sf = new StackTrace().GetFrame(2);
+            var method = sf.GetMethod();
+            var className = method.ReflectedType?.ToString();
+            var methodName = method.Name;
+
+            _initAction?.Invoke();
+            setUp?.Invoke();
+
+            using var context = MockContext.Start(className, methodName);
+            _mockContextAction?.Invoke(context);
+            Helper.SetupEnvironment(AzureModule.AzureResourceManager);
+            SetupAzureContext();
+            var serviceClients = SetupServiceClients(context);
+            AzureSession.Instance.ClientFactory = new MockClientFactory(context, serviceClients);
+            contextAction?.Invoke(context);
+            Helper.RunPowerShellTest(scripts);
+
+            tearDown?.Invoke();
+            _cleanupAction?.Invoke();
+        }
+
+        private string GetTestProjectPath(Assembly testAssembly)
+        {
+            var testProjectPath = testAssembly.GetCustomAttributes<AssemblyMetadataAttribute>().Single(a => a.Key == "TestProjectPath")?.Value;
+
+            if (string.IsNullOrEmpty(testProjectPath))
+            {
+                throw new InvalidOperationException($"Unable to determine the test directory for {testAssembly}");
+            }
+
+            return testProjectPath;
+        }
 
         protected string[] BuildModulesList()
         {
             if (string.IsNullOrEmpty(_callingClassName)
-                && string.IsNullOrEmpty(_newPsScriptFilename))
-                throw new ArgumentNullException($"Both {nameof(_callingClassName)} and {nameof(_newPsScriptFilename)} are null");
+                && string.IsNullOrEmpty(_psTestScript))
+                throw new ArgumentNullException($"Both {nameof(_callingClassName)} and {nameof(_psTestScript)} are null");
 
-            var allScripts = CommonPsScripts;
-            allScripts.Add(_newPsScriptFilename ?? $"{_callingClassName}.ps1");
+            var allScripts = _psCommonScripts;
+            if (_psTestScript != null)
+            {
+                allScripts.Add(_psTestScript);
+            }
 
-            var allScriptsWithPath = _projectSubfolderForTestsName == null
-                ? allScripts
-                : allScripts.Select(s => Path.Combine(_projectSubfolderForTestsName, s));
+            var allScriptsWithPath = allScripts.Select(s => Path.Combine(_testBaseFolderName, s));
 
-            var allModules = RmModules;
+            var allModules = _psAzModules;
             allModules.AddRange(allScriptsWithPath);
 
             return allModules.ToArray();
@@ -264,33 +305,24 @@ namespace Microsoft.Azure.Commands.TestFx
         protected void SetupAzureContext()
         {
             const string tenantIdKey = "TenantId";
-            const string domainKey = "Domain";
             const string subscriptionIdKey = "SubscriptionId";
             const string undefined = "Undefined";
             var zeroGuid = Guid.Empty.ToString();
             const string dummyGuid = "395544B0-BF41-429D-921F-E1CA2252FCF4";
 
             string tenantId =  null;
-            string userDomain = null;
             string subscriptionId = null;
             switch (HttpMockServer.Mode)
             {
                 case HttpRecorderMode.Record:
                     var environment = TestEnvironmentFactory.GetTestEnvironment();
-                    tenantId = environment.Tenant;
-                    userDomain = string.IsNullOrEmpty(environment.UserName)
-                        ? string.Empty
-                        : environment.UserName.Split(new[] { "@" }, StringSplitOptions.RemoveEmptyEntries).Last();
-
+                    tenantId = environment.TenantId;
                     subscriptionId = environment.SubscriptionId;
                     break;
                 case HttpRecorderMode.Playback:
                     tenantId = HttpMockServer.Variables.ContainsKey(tenantIdKey)
                         ? HttpMockServer.Variables[tenantIdKey]
                         : dummyGuid;
-                    userDomain = HttpMockServer.Variables.ContainsKey(domainKey)
-                        ? HttpMockServer.Variables[domainKey]
-                        : "testdomain.onmicrosoft.com";
                     subscriptionId = HttpMockServer.Variables.ContainsKey(subscriptionIdKey)
                         ? HttpMockServer.Variables[subscriptionIdKey]
                         : zeroGuid;
@@ -305,32 +337,41 @@ namespace Microsoft.Azure.Commands.TestFx
             AzureRmProfileProvider.Instance.Profile.DefaultContext.Subscription.Id = subscriptionId ?? undefined;
         }
 
-        protected void SetupMockServerMatcher()
+        protected void SetupMockServer()
         {
-            var resourceProviders = _matcherResourceProviders?.Count > 0
-                ? _matcherResourceProviders
-                : new Dictionary<string, string> // default
-                    {
-                        {"Microsoft.Resources", null},
-                        {"Microsoft.Features", null},
-                        {"Microsoft.Authorization", null},
-                        {"Providers.Test", null},
-                    };
+            var resourceProvidersToIgnore = _matcherResourceProvidersToIgnore?.Count > 0
+                ? _matcherResourceProvidersToIgnore
+                : new Dictionary<string, string>
+                {
+                    {"Microsoft.Resources", null},
+                    {"Microsoft.Features", null},
+                    {"Microsoft.Authorization", null},
+                    {"Providers.Test", null},
+                };
 
-            var extraUserAgentsToIgnore = new Dictionary<string, string> // default
-            {
-                {"Microsoft.Azure.Management.Resources.ResourceManagementClient", "2016-02-01"},
-            };
+            var userAgentsToIgnore = _matcherUserAgentsToIgnore?.Count > 0
+                ? _matcherUserAgentsToIgnore
+                : new Dictionary<string, string>
+                {
+                    {"Microsoft.Azure.Management.Resources.ResourceManagementClient", "2016-02-01"},
+                };
 
-            _matcherExtraUserAgentsToIgnore?.Keys.ForEach(k => extraUserAgentsToIgnore.Add(k, _matcherExtraUserAgentsToIgnore[k])); //extra
-
-            var userAgentsToIgnore = _matcherNewUserAgentsToIgnore?.Count > 0
-                ? _matcherNewUserAgentsToIgnore
-                : extraUserAgentsToIgnore;
-
-            HttpMockServer.Matcher = RecordMatcher(true, resourceProviders, userAgentsToIgnore);
+            _recordMatcher ??= (ignoreResourcesClient, resourceProviders, userAgentsToIgnore) => new PermissiveRecordMatcherWithApiExclusion(ignoreResourcesClient, resourceProviders, userAgentsToIgnore);
+            HttpMockServer.Matcher = _recordMatcher(_matcherIgnoreResourceClient, resourceProvidersToIgnore, userAgentsToIgnore);
+            HttpMockServer.RecordsDirectory = Path.Combine(_testProjectPath, "SessionRecords");
         }
 
-        #endregion
+        protected object[] SetupServiceClients(MockContext context)
+        {
+            if (_initializedServiceClients == null)
+                return Array.Empty<object>();
+
+            var clients = new List<object>();
+            foreach (var client in _initializedServiceClients)
+            {
+                clients.Add(client(context));
+            }
+            return clients.ToArray();
+        }
     }
 }
